restructure code and delete unused
