+++ pkg/types/messaging_peer.go
@@ -151,10 +151,8 @@ func sendBlobs(conn connection.ConnWrapper, ids []uuid.UUID) error {
 			blockCount++
 		}
 
-		_, err = conn.WriteInt(blockCount)
-		if err != nil {
-			return err
-		}
+		conn.WriteInt(blockCount)
+		conn.Flush()
 
 		file, err := blobmngr.FileFromID(id)
 		if err != nil {
@@ -185,42 +183,6 @@ func sendBlobs(conn connection.ConnWrapper, ids []uuid.UUID) error {
 
 		log.Printf("Transfered Blob %s", id.String())
 	}
-
-	return nil
-}
-
-func blobIDsFromMessages(msgs ...Message) []uuid.UUID {
-	ids := make([]uuid.UUID, 0)
-
-	for _, msg := range msgs {
-		if msg.ContainsBlob() {
-			ids = append(ids, msg.Content.Meta.BlobUUID)
-		}
-	}
-
-	return ids
-}
-
-func expectResponse(conn connection.ConnWrapper, expResp string) error {
-	resp, err := conn.ReadString()
-	if err != nil {
-		return err
-	} else if resp != expResp {
-		return fmt.Errorf("received response \"%s\" wanted \"%s\"", resp, expResp)
-	}
-
-	return nil
-}
-
-func fingerprintChallenge(conn connection.ConnWrapper, id Identity) error {
-	challenge, err := conn.ReadBytes(false)
-	if err != nil {
-		return err
-	}
-
-	conn.WriteString(id.Fingerprint())
-	conn.WriteBytes(id.Sign(challenge), false)
-	conn.Flush()
 
 	return nil
 }
