+++ pkg/types/util.go
@@ -3,8 +3,10 @@ package types
 import (
 	"crypto/rand"
 	"encoding/base64"
+	"fmt"
 	"time"
 
+	"github.com/craumix/onionmsg/pkg/sio/connection"
 	"github.com/google/uuid"
 )
 
@@ -41,4 +43,40 @@ func CopySyncMap(m SyncMap) SyncMap{
     }
 
     return cp
+}
+
+func blobIDsFromMessages(msgs ...Message) []uuid.UUID {
+	ids := make([]uuid.UUID, 0)
+
+	for _, msg := range msgs {
+		if msg.ContainsBlob() {
+			ids = append(ids, msg.Content.Meta.BlobUUID)
+		}
+	}
+
+	return ids
+}
+
+func expectResponse(conn connection.ConnWrapper, expResp string) error {
+	resp, err := conn.ReadString()
+	if err != nil {
+		return err
+	} else if resp != expResp {
+		return fmt.Errorf("received response \"%s\" wanted \"%s\"", resp, expResp)
+	}
+
+	return nil
+}
+
+func fingerprintChallenge(conn connection.ConnWrapper, id Identity) error {
+	challenge, err := conn.ReadBytes(false)
+	if err != nil {
+		return err
+	}
+
+	conn.WriteString(id.Fingerprint())
+	conn.WriteBytes(id.Sign(challenge), false)
+	conn.Flush()
+
+	return nil
 }
